requirements.txt
python-decouple
telethon==1.24.0
aioredis
hiredis